requirements.txt
numpy==1.26.0
Flask==3.0.1
flask-cors
openai==0.28
python-dotenv